// libs/ui/pages/admin/Operations/components/PermitDetails/types/permitDetails.types.ts
export interface PermitDetails {
  id: string;
  permitName: string;
  ownerId: string;
  ownerName: string;
  workOrderType: string;
  permitStatus: string;
  permitApprover: string;
  permitApprovalStatus: string;
  permitType: string;
  isActive: boolean;
  durationHours: number;
  associatedEmployeeCount: number;
  componentOfWork: string;
  workLocation: string;
  workOrder: string;
  hazard: string;
  otherHazards: string;
  workDescription: string;
  detailsOfWorkDone: string;
  criteria: string;
  wtg: string;
  currentUser: string;
  ppeRequired: string[];
  approveRejectComment: string;
  winchOperation: boolean;
  isolationRequired: boolean;
  hubLock: boolean;
  lotoPhotosCaptured: boolean;
  hubLockPhotoCaptured: boolean;
  winchPhotoCaptured: boolean;
  properFRPLadder: boolean;
  undergroundOverheadCables: boolean;
  oxymeterNeeded: string;
  firstAidKitNeeded: string;
  handLampNeeded: string;
  state: string;
  area: string;
  permitExpiryAge: string;
  permitRequestedTime: string;
  permitStartTime: string;
  permitValidTill: string;
  extensionApprovalStatus: string;
  extensionGiven: boolean;
  extensionDurationHours: number;
}